Cleaned up stuff from rebase/WIP single product stuff
