+++ Assets/PlayServicesResolver/Editor/play-services-resolver_v1.2.54.0.txt.meta
@@ -1,9 +1,9 @@
 fileFormatVersion: 2
 guid: de05f212ebe54d579a22cb4ad421d36d
 labels:
-- gvh_v1.2.54.0
 - gvh
 - gvh_manifest
+- gvh_version-1.2.54.0
 timeCreated: 1474401009
 licenseType: Pro
 TextScriptImporter:
