--- Assets/Shopify/UIToolkit/Themes/Generic/Prefabs.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 024dd81def5014748ab2575c6c0b46f3
-folderAsset: yes
-timeCreated: 1513106875
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
