--- Assets/Shopify/UIToolkit/Themes/Generic/Prefabs/Dropdown Menu.prefab.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 2d3521fba97d54e079d5593da6d38a26
-timeCreated: 1513179305
-licenseType: Free
-NativeFormatImporter:
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
