--- Assets/Shopify/UIToolkit/Themes/Generic/Prefabs/Selected Option.prefab.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: b3a4f24406d1341b78355c1577da85d7
-timeCreated: 1513179116
-licenseType: Free
-NativeFormatImporter:
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
