--- Assets/Shopify/UIToolkit/Themes/Generic/Prefabs/Unselected Option.prefab.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 5fdf6236689f14b5bb26500c838326d1
-timeCreated: 1513179128
-licenseType: Free
-NativeFormatImporter:
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
